// src/store/enums/async-interaction-type.enum.ts
export enum AsyncInteractionType {
    SAGA_LOADING_SERIES_SEARCH = 'SAGA_LOADING_SERIES',
    SAGA_LOADING_SERIES_INFORMATION_FROM_PORTAL = 'SAGA_LOADING_SERIES_INFORMATION_FROM_PORTAL',
    SAGA_LOADING_SERIES_INFORMATION_FROM_API = 'SAGA_LOADING_SERIES_INFORMATION_FROM_API',
    SAGA_LOADING_SEASON = 'SAGA_LOADING_SEASON',
    SAGA_LOADING_SEASON_FORCED = 'SAGA_LOADING_SEASON_FORCED',
    SAGA_START_EPISODE = 'SAGA_START_EPISODE',
    SAGA_START_NEXT_EPISODE = 'SAGA_START_NEXT_EPISODE',
    SAGA_START_PREVIOUS_EPISODE = 'SAGA_START_PREVIOUS_EPISODE',
    SAGA_START_CONTINUE_AUTOPLAY = 'SAGA_START_CONTINUE_AUTOPLAY',
}